refactor(web): modularize useChat hook into sub-engines to satisfy line limits
